pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "budgt-sh"
version = "0.1.0"
description = "Budgt.sh - A modern Terminal User Interface (TUI) application for tracking personal expenses and budgeting"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Tyler Burr", email = "your-email@example.com" },
]
keywords = [
    "budget",
    "budgeting",
    "expense",
    "tracker",
    "tui",
    "terminal",
    "finance",
    "money",
    "personal-finance",
    "cli",
    "textual",
    "budgt"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Terminals",
    "Topic :: Utilities",
]
dependencies = [
    "textual>=5.3.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "plotext>=5.2.8",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/budgt.sh"
Repository = "https://github.com/yourusername/budgt.sh.git"
Issues = "https://github.com/yourusername/budgt.sh/issues"
Documentation = "https://github.com/yourusername/budgt.sh#readme"

[project.scripts]
budgt = "budgt.app:main"
budgtsh = "budgt.app:main"

[tool.hatch.build.targets.sdist]
include = [
    "/budgt",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["budgt"]

[tool.hatch.build.targets.wheel.shared-data]
"budgt/categories.yaml" = "budgt/categories.yaml"
"budgt/styles.tcss" = "budgt/styles.tcss"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
